Add cocreate action to ideas
The new feature implements a TokBox video-chat box
to a new page called /cocreate. A migration was
added to save a session id that doesn't changes
for each idea after its creation (not update or etc).
Jasmine was also added to handle javascript tests
but it isn't being used yet.